+++ src/tests/_internal/server/conftest.py
@@ -8,6 +8,7 @@ from dstack._internal.server.main import app
 from dstack._internal.server.services import encryption as encryption  # import for side-effect
 from dstack._internal.server.services import logs as logs_services
 from dstack._internal.server.services.docker import ImageConfig, ImageConfigObject
+from dstack._internal.server.services.logs.filelog import FileLogStorage
 from dstack._internal.server.testing.conf import postgres_container, session, test_db  # noqa: F401
 
 
@@ -18,13 +19,11 @@ def client(event_loop):
 
 
 @pytest.fixture
-def test_log_storage(
-    tmp_path: Path, monkeypatch: pytest.MonkeyPatch
-) -> logs_services.FileLogStorage:
+def test_log_storage(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> FileLogStorage:
     root = tmp_path / "test_logs"
     root.mkdir()
-    storage = logs_services.FileLogStorage(root)
-    monkeypatch.setattr(logs_services, "_default_log_storage", storage)
+    storage = FileLogStorage(root)
+    monkeypatch.setattr(logs_services, "_log_storage", storage)
     return storage
